// src/common/enums/project-institutional-state.enum.ts
export enum ProjectInstitutionalState {
  INSTITUTIONAL_IN_PROGRESS = 'INSTITUTIONAL_IN_PROGRESS',
  READY_FOR_PRODUCT_RADICATION = 'READY_FOR_PRODUCT_RADICATION',
  PENDING_PLANNING_RADICATION_CHECK = 'PENDING_PLANNING_RADICATION_CHECK',
  RADICATION_RETURNED_TO_PRODUCT = 'RADICATION_RETURNED_TO_PRODUCT',
  FINALIZED = 'FINALIZED',
}